Decompose bounded test runner
Move bounded process, fixture, and regression-test code behind the project module boundary so the compiler source monolith ratchet remains green without weakening the ceiling.

Refs #1464
